requirements.txt
selenium>=4.9.0
webdriver-manager>=3.8.5
requests>=2.28.2
pandas>=1.5.3
Pillow>=9.4.0
beautifulsoup4>=4.11.2
lxml>=4.9.2
urllib3>=1.26.15
python-dotenv>=1.0.0